Update README.md with setting API version to v3
Certain endpoints for bounces and unsubscribes do not work with the v2 endpoint which the RestClient.php file still uses by default. It appears as though email sending still uses v2 by default too.

I updated the README to explain a workaround for this by setting the API version in the code using $mg->setApiVersion('v3');

A better solution imo would be to update the code to use v3 by default but I am not sure if that would break other things. Please let me know if there is a better solution to this.