Basic settings for GitHub Actions as a replacement for Travis CI
For some reason, Travis CI queues the jobs but takes ages before they start running.
This is, essentially, stopping the development when we rely on the results to move forward.

This GitHub Actions configuration file aims to boot the SDK testing several PHP versions
and running both PHP Unit tests as well as PHP Stan checks.