Fix problem with sending batch email
```
Whenever we use Mailgun API to send emails, and when 'email to' address and 'Reply to' email address are the same, then the recipient variables are not being replaced with the actual values. I have cross checked this with the team and found that Mailgun API is not replacing the notation with actual values in this case.
```